dmaengine/sdxi: Convert sdxi_dev->cxt_lock from spinlock to mutex
The SDXI context initialization and teardown code performs sleeping
allocations with the sdxi_dev->cxt_lock spinlock held. This is a bug:

 SDXI: Device 0x0801 found [cap0=0x816000801, cap1=0x18003f8954]
 BUG: sleeping function called from invalid context at include/linux/sched/mm.h:306
 in_atomic(): 1, irqs_disabled(): 1, non_block: 0, pid: 9, name: kworker/0:1
 ...
 Call Trace:
  dump_stack_lvl
  __might_resched
  __kmem_cache_alloc_node
  kmalloc_trace
  alloc_cxt
  sdxi_cxt_alloc
  sdxi_cxt_init
  sdxi_device_start
  sdxi_pci_probe
  local_pci_probe
  work_for_cpu_fn
  process_one_work
  worker_thread
  kthread

Since this lock is only used in process context it can be changed to a
mutex without restructuring the code. If for some reason this actually
needs to be a spinlock then the allocations will need to be moved out
of the critical sections, which will be a bigger project.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>