dmaengine/sdxi: Rename error handling labels in sdxi_sq_alloc()
It's much easier to understand code that uses local labels for error
recovery/unwinding when the labels describe what they actually do
instead of the condition they're recovering from.

This function currently leaks DMA mappings in its error paths, so it
needs more labels. Rename the existing labels so that the fix to
follow is easier to review.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>