dmaengine: sdxi: New internal header for architected types
Most of the structs in the sdxi driver that represent the memory
layout of data structures defined in the SDXI spec use C bitfields and
don't define the byte order, for example:

  struct cxt_l2_entry {
          u64 vl                  : 1;
          u64 rsvd                : 11;
          u64 l1_ptr              : 52;
  };

The layout of bitfields is implementation-defined and they basically
should not be used for types at an ABI boundary. And if a big-endian
arch ever wanted to use SDXI, new versions of the types would have to
be defined.

We're going to convert all of these to structs with explict byte order
accessed via mask and shift, starting with the level 2 context table.

No functional change, just define the types and a couple accessors.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>