dmaengine: sdxi: Avoid bitfield use in MMIO_ERR_CTL setup
Just define the bit for the interrupt enable flag and use it directly,
allowing us to remove union mmio_err_ctl_reg.

Minor functional change: the err_ctl_reg temporary was uninitialized
except for the .en bit. This change ensures that all the reserved bits
in the register update are zeroed.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>