dmaengine: sdxi: Stop devices before initialization
An SDXI device can be in arbitrary state when it's probed. The code
currently assumes it's in GSV_STOP, but that's not always the
case (kexec, for example).

The spec requires the OS to stop the function before modifying
MMIO_CXT_L2.

Drive the device to stopped state before proceeding with its
configuration. We might consider logging when we find the device in an
"interesting" state (i.e. anything that isn't GSV_STOP).

The new code introduced here should probably live in device.c?

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>