dmaengine: sdxi: Initialize dma_device->directions correctly
The @directions member of struct dma_device is a bit field; initialize
it accordingly. (Setting it to DMA_MEM_TO_MEM just sets it to zero.)

I don't think this really results any functional change for SDXI since
dma_get_slave_caps() is the only code that consults the field and it
should already return -ENXIO for SDXI channels.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>