dmaengine: sdxi: Stop all contexts on device exit
We have code that checks for active contexts at device exit. Let's
just convert it to stop any contexts it finds now that we have the
code to do it correctly. Note that we must stop the admin context
last (should we be issuing a stop command to it? I guess we'll find
out).

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>