dmaengine: sdxi: Follow spec for function activation
From 4.1.8 "Activation of the SDXI Function by Software"
  ...
  9. Set MMIO_CTL0.fn_gsr to GSRV_ACTIVE.
    a. Wait until MMIO_STS0.fn_gsv is GSV_ACTIVE or GSV_ERROR before
       performing any other actions on the function.

The current code is just setting MMIO_CTL0.fn_gsr and proceeding
without waiting for MMIO_STS0.fn_gsv to reflect the state
transition (whether to active or error).

Introduce a convenience function to perform the activation by the
book. Add the now-necessary error handling, since starting the device
is fallible. Remove a debug statement that's in the middle of all this
and not that useful (if we need to report the DMA addresses of control
structures we can do that in a nicer way later).

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>