iommu/amd: Enable PASID Privileged Mode when supported
Enable PASID Privileged Mode when supported to allow endpoints to send
requests with the Privileged Mode Requested bit set. While the meaning
of Privileged Mode depends on the protection model of the system and
outside the scope of the PCIe spec, the intention is to allow the
system to distinguish between requests that target memory belonging to
a supervisory entity such as a kernel or hypervisor and requests
targeting memory belonging to less privileged entities such as user
space processes.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>