dmaengine: sdxi: device: Improve error state handling during stop
I occasionally see functions stuck in error state on driver attach on
Turin but the condition is hard to recreate and debug. This is an
attempt to address or mitigate after consulting the spec.

Here is what SDXI 1.0 says about how to recover a function from error
state (4.1.6 G5: GSV_ERROR State):

"""
  Software may return the function to the GSV_STOP state by performing
  one these actions.

    1. Set MMIO_CTL0.fn_gsr to GSRV_RESET and wait until
       MMIO_STS0.fn_gsv is GSV_STOP. [....]
    2. Perform a function-level reset at the device level.
"""

I think (1) is the only viable option for the sdxi driver;
attempting (2) during probe seems too disruptive/difficult to me right
now.

Note that the spec says "set MMIO_CTL0.fn_gsr to GSRV_RESET *and wait*
until MMIO_STS0.fn_gsv is GSV_STOP". The current code for driving the
function to the stopped state diverges from this in that
it *repeatedly* issues reset commands to the function until it gets
SDXI_GSV_STOP or hits the timeout. At least on my Turin/Quartz
testbed, this fails to transition the function to stopped state and
times out. (I did not save logs, unfortunately.)

Tease apart the questionably combined SDXI_GSV_STOP and SDXI_GSV_ERROR
case in sdxi_dev_stop(), ensuring that we issue a single reset command
and we only do it when the function is in error state. There's no
reason to do it when it's stopped. Also don't discard any other state
in MMIO_CTL0.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>