dmaengine: sdxi: Remove IOMMU_SVA Kconfig dependency
The whole of the driver should not be gated by IOMMU_SVA. The DMA
engine provider and core support code are useful without it. With this
change the character device is still created but is useless; maybe
that should be changed in future commits by making it a separate
module.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>