[INTERNAL] selftests/sdxi: Add pteditor module and library
In order to enable validation of SDXI hardware behavior with respect
to PASID Privileged Mode, vendor the MIT-licensed pteditor module:

  https://github.com/misc0110/PTEditor

as of commit e846ff0d98d8 ("remove old code" - Apr 24 2025).

This module allows unprivileged manipulation of page tables and is a
security hole. It's for testing only and should never be used in a
production setting.