[INTERNAL] selftests/sdxi: Add PASID Privileged Mode verification
Verify that the system allows SDXI copy operations to access data
buffers in arbitrary memory when the akey has the 'pr' bit set, and
that the system denies such accesses to supervisor memory when the
akey 'pr' bit is unset.

This test expects to be able to manipulate the following sdxi module
parameters:

  set_pr_bits: Set the 'pr' bits on kernel-private SDXI control
  structures when the underlying bus supports privileged address space
  and the function has been configured to use it (e.g. PCIe PASID
  Privileged Mode) (default: false) (bool)

  force_pr_for_user_contexts: Force-enable the 'pr' bit for user
  contexts. Not useful without set_pr_bits=1. This is a security hole
  and is intended for hardware validation only. (default:
  false) (bool)

It also requires the test-specific "pteditor" module to be loaded.