dmaengine: sdxi: Add sdxi_akey_table type and akey_table_order() helper
Introduce a type for an akey table whose size is explicitly set to the
minimum (4KB) at build time. This ought to be fine to use until we
encounter a need to dynamically size the akey table at context
creation.

Since we have to encode the size of the table in the context L1
entry, introduce a helper function to report a table's order. This can
be be fleshed out later for tables > 4KB when needed.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>