dmaengine: sdxi: Convert sdxi_sq->csb to coherent DMA allocation
Use dma_alloc_coherent() and friends to allocate and map the
completion status array instead of using the streaming DMA API, which
is intended for short-lived single transfers.

Ultimately we may want make descriptor submission opportunistically
allocate completion status blocks from a per-context pool, but this is
the straightforward conversion.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>