dmaengine: sdxi: Use coherent DMA allocations for L1 tables
This seems to be the least mess I can make of this conversion.
set_cxt_l2_entry() now only takes the DMA address of the L1 table, and
it is for use only in the setup path. The L2 entry is cleared on
teardown in clear_cxt_table_entries() now. L1 table allocation and
mapping is now performed in confix_cxt_tables().

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>