dmaengine: sdxi: Test sdxi_encode_size32()
Introduce a single test case for sdxi_encode_size32() that checks a
variety of inputs directly without involving a descriptor.

This could have been a table-driven parameterized test, but the number
of inputs that reasonably need to be checked is small, and it would
introduce a bit too much indirection to justify the readability
penalty in this case.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>