dmaengine: sdxi: descriptor: Expect vl=0 after encode in tests
To support serializing descriptors directly to the hardware ring, we
need more precise control over when they are marked valid. Therefore
the encoder functions will change to leave the vl bit clear;
higher-level code is expected to set vl later, when it really wants
the descriptor to be consumed.

Update the tests to reflect this (this makes them fail, as expected).

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>