dmaengine: sdxi: dma: More implementation
Add a completion status block to struct sdxi_dma_desc that will signal
the completion of the final SDXI descriptor in each virtual
descriptor. This is allocated and initialized but not yet associated
with the descriptor at the SDXI level. For now we force the error bit on
so all transactions fail.

Implement dma_device->device_tx_status() and ->device_synchronize().

A polled mode dmatest run now finishes without warnings or oopses;
naturally all descriptors fail.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>