dmaengine: sdxi: dma: Improve terminate_all and synchronize
Nop out all the pending descriptors that have yet to be issued in
sdxi_dma_terminate_all(). As long as all valid bit updates are
performed under the channel lock, this ought to be safe.

sdxi_dma_synchronize() submits a fenced nop and busy-polls for
completion. (It should eventually be changed to submit an intr
descriptor and wait.)

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>