dmaengine: sdxi: descriptor: Interrupt mitigation
Write a DSC_INTR to the ring only at issue time. The prepare step
leaves a nop after each copy; the issue overwrites the last nop in the
submitted list with an interrupt.

This achieves some interrupt reduction with dmatest and multiple
threads per channel, but not a lot, and it's ugly. Committing it now
so I have it in the history but I'll probably revert it.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>