dmaengine: sdxi: Remove enqueue demonstration code
sdxi_enqueue() and associated code were adapted from the SDXI spec.
The intent of this code is to demonstrate lockless submission of SDXI
descriptors to the ring. While it accomplishes that goal, it's not
suited to how we want to manage the descriptor ring for in-kernel
clients. The ring state and reservation APIs are more flexible and
do not require copying of whole descriptors.

Signed-off-by: Nathan Lynch <nathan.lynch@amd.com>